% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myadd.R
\name{myadd}
\alias{myadd}
\title{Adding two numbers}
\usage{
myadd(a, b)
}
\arguments{
\item{a}{the first number to be added}

\item{b}{the second number to be added}
}
\value{
the sum of the two input values
}
\description{
This function just adds two numbers. It returns the result.
Next line illustrating that this can be multiline.
}
\details{
This function adds two numbers.
}
\author{
Csaba Farago
}
