% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymultiply.R
\name{mymultiply}
\alias{mymultiply}
\title{Multiplying two numbers}
\usage{
mymultiply(a, b)
}
\arguments{
\item{a}{the first number to be multiplied}

\item{b}{the second number to be multiplied}
}
\value{
the prod of the two input values
}
\description{
This function just multiplies two numbers. It returns the result.
}
\details{
This function multiplies two numbers.
}
\author{
Csaba Farago
}
